+++ packages/terra-date-time-picker/tests/nightwatch/components/DateTimePickerDisabled.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from './DateTimePickerExampleTemplate';
+
+const DateTimePickerExample = () => (
+  <DateTimePickerExampleTemplate
+    disabled
+  />
+);
+
+export default DateTimePickerExample;