+++ packages/terra-time-input/src/TimeInput.jsx
@@ -12,6 +12,10 @@ const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
+   * Whether the time input should be disabled.
+   */
+  disabled: PropTypes.bool,
+  /**
    * Custom input attributes that apply to both the hour and minute inputs.
    */
   inputAttributes: PropTypes.object,
@@ -47,6 +51,7 @@ const propTypes = {
 };
 
 const defaultProps = {
+  disabled: false,
   inputAttributes: {},
   minuteAttributes: {},
   hourAttributes: {},
@@ -492,6 +497,7 @@ class TimeInput extends React.Component {
 
   mobileInput() {
     const {
+      disabled,
       inputAttributes,
       minuteAttributes,
       hourAttributes,
@@ -549,6 +555,7 @@ class TimeInput extends React.Component {
             onBlur={this.handleHourBlur}
             size="2"
             pattern="\d*"
+            disabled={disabled}
           />
           <label htmlFor={hourAttributes.id} className={cx('mobile-input-label')}>
             {this.context.intl.formatMessage({ id: 'Terra.timeInput.hours' })}
@@ -571,24 +578,25 @@ class TimeInput extends React.Component {
             onBlur={this.handleMinuteBlur}
             size="2"
             pattern="\d*"
+            disabled={disabled}
           />
           <label htmlFor={minuteAttributes.id} className={cx('mobile-input-label')}>
             {this.context.intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
           </label>
         </div>
         {this.props.variant === TimeUtil.FORMAT_12_HOUR && (
-          <ButtonGroup isSelectable onChange={this.handleMeridiemButtonChange}>
+          <ButtonGroup selectedKeys={[this.state.meridiem]} onChange={this.handleMeridiemButtonChange}>
             <ButtonGroup.Button
-              key="anti_meridiem_button"
+              key={this.anteMeridiem}
               className={cx('meridiem-button')}
               text={this.anteMeridiem}
-              isSelected={this.state.meridiem === this.anteMeridiem}
+              isDisabled={disabled}
             />
             <ButtonGroup.Button
-              key="post_meridiem_button"
+              key={this.postMeridiem}
               className={cx('meridiem-button')}
               text={this.postMeridiem}
-              isSelected={this.state.meridiem === this.postMeridiem}
+              isDisabled={disabled}
             />
           </ButtonGroup>
         )}
@@ -596,12 +604,13 @@ class TimeInput extends React.Component {
     );
   }
 
-  handleMeridiemButtonChange(event, selectedIndex) {
-    this.handleValueChange(event, TimeUtil.inputType.HOUR, this.state.hour.toString(), selectedIndex === 0 ? this.anteMeridiem : this.postMeridiem);
+  handleMeridiemButtonChange(event, selectedKey) {
+    this.handleValueChange(event, TimeUtil.inputType.HOUR, this.state.hour.toString(), selectedKey);
   }
 
   desktopInput() {
     const {
+      disabled,
       inputAttributes,
       minuteAttributes,
       hourAttributes,
@@ -661,6 +670,7 @@ class TimeInput extends React.Component {
           onBlur={this.handleHourBlur}
           size="2"
           pattern="\d*"
+          disabled={disabled}
         />
         <span className={cx('time-spacer')}>:</span>
         <Input
@@ -679,6 +689,7 @@ class TimeInput extends React.Component {
           onBlur={this.handleMinuteBlur}
           size="2"
           pattern="\d*"
+          disabled={disabled}
         />
         {this.props.variant === TimeUtil.FORMAT_12_HOUR && (
           [
@@ -691,6 +702,7 @@ class TimeInput extends React.Component {
               value={this.state.meridiem}
               size={this.state.meridiem.length || 1}
               readOnly
+              disabled={disabled}
             />,
             <div
               style={{ position: 'relative' }}