+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+*Added disabled prop.
+
+### Changed
+* Uplift to Button 2.0
 
 1.7.0 - (February 1, 2018)
 ------------------