+++ packages/terra-toggle-button/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Uplift to Button 2.0
 
 1.19.0 - (February 1, 2018)
 ------------------